first changes to make f03gl work

+++ f03gl/OpenGL_gl.f90
@@ -550,7 +550,7 @@ INTEGER(GLenum), PARAMETER :: GL_REPEAT                = z'2901' ! 0x2901
 INTEGER(GLbitfield), PARAMETER :: GL_CLIENT_PIXEL_STORE_BIT = z'00000001' ! 0x00000001
 INTEGER(GLbitfield), PARAMETER :: GL_CLIENT_VERTEX_ARRAY_BIT = z'00000002' ! 0x00000002
 INTEGER(GLbitfield), PARAMETER :: GL_CLIENT_ALL_ATTRIB_BITS =                 &
-    transfer(z'ffffffff',GL_CURRENT_BIT) ! 0xffffffff
+    transfer(INT(z'ffffffff'),GL_CURRENT_BIT) ! 0xffffffff
 !  polygon_offset values
 INTEGER(GLenum), PARAMETER :: GL_POLYGON_OFFSET_FACTOR = z'8038' ! 0x8038
 INTEGER(GLenum), PARAMETER :: GL_POLYGON_OFFSET_UNITS  = z'2A00' ! 0x2A00
